Broke out tests into functional and behavioral
